refactor(dropdown): bring dropdownButton variant prop to child class
