+++ src/components/Dropdown/index.ts
@@ -1,3 +1,2 @@
 export * from "./sgds-dropdown";
 export * from "./sgds-dropdown-item";
-export type { DropdownButtonVariant, DropDirection } from "../../base/dropdown-element";
